# internal/handler/get_subs_by_user/handler.go
package get_subs_by_user

import (
	"net/http"

	"github.com/4udiwe/subscription-service/internal/entity"
	h "github.com/4udiwe/subscription-service/internal/handler"
	"github.com/4udiwe/subscription-service/internal/handler/decorator"
	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
	"github.com/samber/lo"
)

type handler struct {
	s SubscriptionService
}

func New(s SubscriptionService) h.Handler {
	return decorator.NewBindAndValidateDecorator(&handler{s: s})
}

type Request struct {
	UserID uuid.UUID `json:"user_id" validate:"required,uuid"`
}

type Response struct {
	Subscriptions []Subscription `json:"subscriptions"`
}

type Subscription struct {
	SubscriptionID uuid.UUID `json:"subscription_id"`
	UserID         uuid.UUID `json:"user_id"`
	OfferName      string    `json:"offer_name"`
	Price          int       `json:"price"`
	StartDate      string    `json:"start_date"`
	EndDate        string    `json:"end_date"`
}

// Get all subscriptions by user ID
// @Summary Получение подписок по ID пользователя
// @Description Получение списка подписок для указанного пользователя
// @Tags subscriptions
// @Accept json
// @Produce json
// @Param user body Request true "user ID"
// @Success 200 {object} Response
// @Failure 500 {string} ErrorResponse
// @Router /subscriptions/by-user [post]
func (h *handler) Handle(c echo.Context, in Request) error {
	sub, err := h.s.GetAllSubscriptionsByUserID(c.Request().Context(), in.UserID)

	if err != nil {
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}

	return c.JSON(http.StatusOK, Response{
		Subscriptions: lo.Map(sub, func(s entity.SubscriptionFullInfo, _ int) Subscription {
			return Subscription{
				SubscriptionID: s.ID,
				UserID:         s.UserID,
				OfferName:      s.OfferName,
				Price:          s.Price,
				StartDate:      s.StartDate.Format("2006-01-02"),
				EndDate:        s.EndDate.Format("2006-01-02"),
			}
		}),
	})
}